changing configure to automatically choose 64 bit or 32 bit compilation - fixed ikrand to make this possible --Molly
git-svn-id: https://giordano.mit.edu/nvs/mangle/trunk/dev_80bit@291 9f78a35e-8a50-488b-8db1-b351ce30c53d

+++ src/ikrand.c
@@ -27,9 +27,9 @@ void ikrand_(int *ik, double *ikran)
     /* generate pseudo-random unsigned long long */
     } else {
 	llikran = (unsigned long long *)ikran;
-	*llikran = random();
-	*llikran <<= (8 * sizeof(unsigned long));
-	*llikran |= random();
+	*llikran = (unsigned int)random();
+	*llikran <<= (8 * sizeof(unsigned int));
+	*llikran |= (unsigned int)random();
     }
 }
